python_backend/requirement.txt
opencv-python
langchain
langchain-openai
langchain_google_genai
langchain-community
python-dotenv
pyaudio
soundfile
SpeechRecognition
SpeechRecognition[whisper-local]
plyer
wikipedia
pyautogui
pyttsx3
install mpg123 using choco
flask
flask_core
psutil
whisper
pygame
wave
TTS
geocoder
signal
pycaw 
comtypes
screen-brightness-control
google-generativeai
streamlit
PyPDF2
chromadb
faiss-cpu
